Establish an initial state for bindery tracking
This starts pretty deep into the project.
AUTH is in place
ACL is being used but may be overkill
Images are being tracked but this may be a usefull as a branch
Product galleries are operating at a basic level meaning,
  the thumbnails page and navigate
  the gallery detail pages come up and format properly
But
  Admin support for updates doesn't exist yet.
News/Dispatch is not yet working, meaning,
  There is not yet sufficient data to write the thumbnail pulls
  Nor to create the On The Bench pages (nor to decide on desing)
But
  Reasonable progress has been made toward creating admin
  update tools for Dispatches.
  These tools may translate as a core for handling Product/Exhibit support
Basic site navigation seems stable and shouldn't need much more work until
a later phase. Problems that may turn up: excessive length given the
left hand real estate relative to the number of choices