Wrote a little next/prev clicker to use in standard crud environments.
Whacked on the dispatch/ingest_image system. Got it working... sort of. It managed to process thumbnails and save image and dispatch records. But it wouldn't save the link between those records. Then the next morning it wouldn't work. It objected to the way the img_file field was offered (as an array, like it would in a normal upload situation).

I think I've mess with that system for long enough. I'm going to try working from inside Image and repurpose the folder analysis tools (disallowed, duplicate, new) that I used in dispatch/ingest_image. I'd like to get away from all the ajax and make things a little simple and more direct.

The last change was to beginings of an image-grid presentation. This first simple version pulls ALL the images and provides a simple form on the page so the number of columns and size of the images can be specified.