Just chasing a minor clean up along with the last commit.
There had been a ton of directory reference problems in image uploading that had to be fixed and this one related to making the picture destination directory flexible rather than fixed. I had tried it in several places but eventually found it had to be in the behavior because the property was protected.

So this got rid of a now-unused attempt.