Cobble together a fix for the broken search on Gallery that is triggered with a menu selection.
Line 71 of gallery had condition 'Gallery.label like' => '%param_thingy%' and I'd deleted label. That field had a lower case, one word and unique version of  Gallery.heading (which is the dispayed menu text) so I just re-conditioned the search and things hooked back up. But this is a temporary fix I think.