The data tables have been modified to bring all images into control of the Image table and to bring Dispatch and Exhibit into the Content table.
The data has been translated into the new tables and all the image specific links have been transfered and all the actual image files moved. A couple of code changes have been made so that the site use (with the exception of normal crud layouts which will be re-spawned) of images is operational under the new scheme.

All the Controllers and Models still exist because there is a lot of work to do to get the Gallery system transfered into Collections. The redundant code will have to stay around for a while yet.