This is a huge commit that incorporates all the schema changes (shifting to Image-Content-Collection).
This rebuilds product galleries to basic functionality under the new streamlined schema. 

And it creates News Feeds to basic function too. In fact, news feeds have the beginning nucleus of an in-place editing system (though it needs to be fleshed out.

The old Exhibit-Dispatch components are still in the code. After Image upload is resolved that legacy code won't have anything to contribute and could be deleted.

The new page layout which was designed for news feeds has been created though it's only hooked up to a few pages. Also, a no-filmstrip version has been spawned and is starting to be used in a few places.

Routing was changed to come in line with the new data structure. It's a bit simpler than the previous arrangement though that code is also still in the file (commented out). I learned enough about routing to re-institute an Admin parent menu to contain a lot of back-end menu choices. This isn't 100% yet. In a few cases the url has to have the admin/ removed (ACL for example). And there may be problems with the user/login paths.

Work has started on the Image Grid backend interface. Mostly this centers around getting a search feature hooked up and that is the last major site change.

A basic plan for Site Search has been laid out and the search box has become live (though there are many missing methods... in fact the majority is missing).