This commit continues to develop Image backend tools.
The image controller now watches the Upload folder, and presents links and tools to process uploaded images. This code is 90% functional. The process for handling disallowed images is the holdout area.

Though these notes aren't long, there was a lot of functionality added in this commit. A lot of efficiency was built into the code through the heavy use of the beforeFilter() in images. Essentially, when the Image controller runs, a large kit of data is assembled for both the Image methods and the Views. Also, Sessions make the Image system the most state-aware area of the site. Screen formatting choices and search sets are remembered as the images are managed.

The plan is to make the Image Grid a center of Content/Collection assembly but this is the least developed part of the image system.

Dispatch, Exhibit and Gallery models and controllers are still in place but it's just about time to dump this baggage. The Admin router changes are still in place but it's looking like a less than ideal solution since it's interfering with User and Home routes.