A couple of minor Model link corrections.
Serious work exploring a Form Element strategy to make forms modular, portable and easy to incorporate in different layouts. This first effort is a bit of a mess. Workable but not sustainable or extensible. This commit is to establish a baseline to begin improvements. First up—adding a Helper to simplify things.