This lays some groundwork for an Image CleanUp routine. This is a layout devoted to the deletion of Image records, Image files and records linking Image records up to Collections.
Image records and the image files can (theoretically) be deleted (or not) at this point.

The interface presentation is being expanded to show the dependent records and present the tools for their deletion. Supporting controller code for these deeper deletions is not in place.

The view output code is a fiasco! This could be massively simplified.