This fills out the Image CleanUp layout to display all the dependent records and checkboxes to control the deletions. Still no deep support for dependent record deletions but image record and image file deletion is working fine. I didn't do anything to simplify the View code except space it out a bit and add a few comments.
The other work in this commit: New Image upload processing. I cleared out some errors that had cropped up on the layout and tweaked the forms to pre-load data properly. I've actually tested this form now and have added a few new images to the system through it.

The fieldset for bulk-uploads is still the old hard-coded element rather than the array and helper version, but it is working properly.