Rebuild action/search logic so different views can more easily share search sets.
Only image_grid and clean_up are using the system right now but that's as far as I'd gotten before making this change. 

There's a bit of polish to put on this probably but it seems to be working correctly.

Next to fold into the pattern: index and multi_add. But index uses the paginator and multi_add needs a View change to display using the data array that will be provided.