Preliminary noodling exploring a Art section strategy. Can't get the url to output properly (keeps inserting /contents) but doesn't really break anything either (at least not anything that wasn't previously broken)
This was an attempt to pluck off the last param[pass] and plug it in as param[pname] then hijack the standard content/gallery() action for output.

It did pull the collection. it's just that pesky route writing problem. I could take all the array elements that are passed to the link() helper and string-ize them. That big hammer should work.