Step One: Fat Model Workshop Module
This is some starter code for a fat model approach. It uses the __construct() method of the model (all classes will run this magic method when they instantiate if it exists).

In this style, when the model starts up, it will run some finds and store the results in some properties. In the controller then, all we have to do is move the contents of the properties to variables that the View can see.

The basic View uses debug to see the passed data.