Fix missing id problem on newsfeed page admin-edit forms
The data array that populated the page and the forms used content_id as a field name. Of course id would be expected. So I changed the array that gets delivered, then chased down the scattered references to the old name.

Also I stopped the garbage that was being dumped (and invisible) in the Jump Box div. The Jump Box tool still doesn't work, but the massive amount of garbage output has been eliminated.