Fix the click and highlight behavior of the newsfeed filmstrip
Repaired a wide range of problems including the failure of prev/next page/image clickers. Almost the only thing that was working was direct thumbnail clicks. Not it's all working.

Also, the jumpbox should work for newsfeeds if it just opens a new page with a proper url (see the issue 13 in github)