Build out the JumpBox feature for Gallery and Newsfeed pages
A bit of javascript to validate the form, expanding the view to place more data in the form, and controller code to construct the redirect url's for the jumps. Amazingly, building the right url and redirecting to it is all that's required for the feature!

Later I'll have to write jumps for other contexts.