Integrate Supplement, Category and Collection.slug
With the new Supplement table in place I had to dig down into the queries that pull content records. But the bigger change came when I put Category in play to normalize Collection categorization and Supplement selection. This cascaded into tackling the problem of multi-word Collection headings in URLs.

To resolve that bad practice I made Collection sluggable and updated the data set. Now I have to do extensive work to fix all the menus and other link-generating code so they use Collection.slug instead of Collection.heading.

And I need to continue on the work of pulling the Collection record of the correct Category as well as getting the Supplement data required for output.

Serious and deep changes. And it all happens here!