Change links so their built on slugs rather than headings
This required a wide scattering of fixes including the installation of the Sluggable behavior on the Collection Model and changes to Navline data

In fact, though I thought the Collection.slug field would be the source for url construction, the panes are really hand written into Navline records. Not such a good thing. Possibly a problem later.