Convert blogs to more specific url
simplify the blog route while simultaneously making the url collection-aware. This serves as the basis for filtering out duplicate Content records under the same heading in different collection.

The blog() action was made much smarter too. It can figure out a collection to use even if the url is pretty unspecific.

Links and menus to blogs were populated with collection id's too.