Advance "Product" landing page with better "Recent News" data
The major change here is a sophisticated query to get recent news entries.

This version will get a specific number of entries, each a with a unique heading, and each the most current of its kind, filtered to the 'dispatch' category.

The view starts to rough out some links using this new data set but is far from it's final format.