Change link method for Supplement records
Supplements were double linked to Image and Collection so they could be unique to a context. However, ContentCollection is a single record that identifies that unique use. So I switched the models to use this simpler logic.

Also baked fresh CRUD for Supplements and did some beginning hack work to bend Controller/View logic over to deal with the new Model logic. More to do there and a more complete way of recording default supplement data sets for a Collection Category is planned.