Established basic functionality for Galley editing
This gets some form fieldsets that toggle and same.

Not all data is available on the form yet. And when the Supplement data is included, changes must be reflected on the page immediately through javascript.

For phase 1.5, some improvements can be made:

If some tables don't save, the entered data will be lost

If some tables aren't changed, they will re-save anyway