Part 1: Adding Supplements to Exhibit Editing
This builds out a rough fieldset element for Supplements.

This is a fall-back position before I try to create a Helper Class inside the Element.

The tasks for this one are pretty complex and I started adding functions to break things up. Then the functions started needing to move variables in and out of themselves and…

Going to a Class might help this. Besides, this Element needs to make some Form HTML and that requires the Form Helper (which is not available in an Element).

So this Class is going to extend FormHelper… I hope.