Install Collections menu
This brings in the collapsed Collections menu with all the click functions to open and close sections all the way through selecting an article.

no auto-opened sections or 'active' highlighting. I'm not sure what the plan might be here.

Some css improvements were made to this new menu section.

Also a 'related' button is added with the detailed articles and the style is set. This needs javascript to operate though.