eliminate Content.publish use
This marks the shift away from any use of Content level publish/hold indication. ContentCollection and Collection level filtering must carry the load.

Whether this took care of all problems, we'll find out in use. The field has been renamed but not removed. The pieces of code have been REM'd but not removed.