Convert image_grid to multi record save - Phase 2
This changes the view so all the record data is output in a single form so the single submit button sends them all back for processing.

There is one small css fix to get the white backgrounds back after moving the form context-relationship.

Lastly, javascript was added to watch for changes in every single field. A new field was added to each record block that gets set from 0 to 1 if a field in the block changes. On the server side we can filter on this and only save the records with changes. I could extend this to Model level tracking too but this will get the job done.

The controller dies when it gets the data. Next phase is to write the loop for the save calls.