Create Art landing page system
If no art gallery entries are found, children of the menu entry are polled for links to construct a landing page.

appHelper got a work-over to reduce redundant code in the search and landing block methods. There is still a lot of overlap but it's better and good enough for now.