develop change_collection()
add and increment/decrement flag to track changed fields for a record.

0 = no fields changed
 n > 0 number of fields changed

Next, when a field changes it gets flagged and receives a 'reset' feature. This should only happen if it doesn't already have those characteristics. As it is now, I count not only them number of changed fields but how many times a given field gets changed!