Continue refining new_upload for Lighroom strategy
hold and delete process now work on the incoming images

all images are placed in holding folders rather than just the native images.

Upload behavior needed to be turned off before any of the necessary tasks for layout generation happened, but that would be bad. It kept destroying the value in img_file so an alternate field holds the file name temporarily. Before the actual save I'll turn off the Behavior and move the data to the correct field.