Begin developing art page styles and page behaviors
Create a starting Collection description style in relation to the gallery image and create the show/hide behavior for the description.

On first entry (no page:) show the description. At all other times hide it but always have a toggle on it.

Also if javascript is off show everything and don't have overlapping elements.