set up foundations for materials page
beginning with the components I developed in the javascript class, install them in the cake framework.

This was a pure javascript solutions with a straight html page and db query data that was written as son into the page for the javascript's use.

I'll want a very different system now. Possibly a simple grid-view page of swatches for non-javascript visitors and all the interactive features layered on later in the browser.