Correct problem with incorrectly unique Content.slug
the sluggable behavior doesn't allow duplicates. I re-wrote it to allow these as a basis for article accumulation. But for a freshly created record there was no way to allow dups when appropriate and force uniqueness when appropriate. So after a content record is created, it is reserved with it's new id and slug calculates and resaves as it should.