Expand and refine pay page. Bring more content to page.
Linked Catalog and Collection so Catalog pages can have context data. Also, Catalogs will now be available from other parts of the site (but I'm not decided on the Collection Category yet).

other files had whitespaces removed because of a failure  (which was resolved when a different strategy was implemented).

Output, css and js behaviors were all refined. The table filter behavior is now an opacity change.