Drag materials selection tool onto purchase pages
It's kicking. It's screaming. But it's there and it's staying.

broke the table collapsing in the process. Possibly the materials page javascript whacked something? That makes no sense…

I think there are some wrappers around the materials page that are unnecessary. Some element clean up is in order but this will probably break the formatting.

Product selection will now need to set the conditions on the materials selector. And the selectors tools will need to be hidden or removed.

Still, all in all, a big step.