fix errors that displayed in new workshop detail code
I reinstated the second part of the condition for the workshop id query and created the missing records in workshop to support the landing pages.

This still left the error though.

Then I saw the query, which was only supposed to return 1 field was digging deep into the relationships. So I added another parameter:

'contain' => false

to stop the related data search. That too care of the error.