prevent obscure error if nameSpace contains the string 'master'
Contact page provides testing platform for the javascript.

Fix one comment

Fix the selector and regex so the only option attributes with the prefix "master" or "master-" will be detected.

This prevents option attributes that are slaves, in nameSpaces with the string 'master' in them from being incorrectly treated as master nodes.