bring option-set manager near to completion
build the initializer that hides all slave for all masters on page load.

work out the hide and show processes for the click method.

Write up more test lines in the contact page.

Having slaves that are also masters complicates the page. The slave of a slave/master must be tagged so it will respond to it's master's master also. Otherwise it might be left showing inappropriately.