New Option element and evolve purchase pages to have options for each product table
this is a rough approximation. There will need to be logic which can discover which options belong to which products

This also adds the new Titling Option element to the set of options.

This element tries to make use of the new toggling-set javascript. But the Fieldset helper is not sophisticated enough to really work with this code properly.

It does work without error, and the slave elements do respond, but the Fieldset helper doesn't allow the super-fine-grain application of element attributes that is required in highly complex cases.

For example, I can make a radio button set act as a master, but I can't make the individual radio buttons control different slave-set members… I'll double check input-radio helper code to confirm that.