start refining master/slave implementation for product purchase tables
Improve the Fieldset helper so an array of attributes can be passed in for the fieldset tag. This is necessary since these tags will need to be slaves to the product choices

tweak up the options_titling element so the radio buttons act as masters controlling the other fields in the fieldset.