install tools for returning data to controller from purchase page
and to debug the returned array for planning.

One problem is how to detect the options that are appropriate to the product. These are filtered to visibility, but that means nothing to the submit button.

I could modify the master/slave function to handle some setting to help, but this would make that very generic tool very specific to forms…

Probably better is to set the submit button to do some processing before it sends the data to the server.