Pause point in add-to-cart ajax
Filtering to only valid fields now works.

Call to server and return that proves success is in place.

The problem is, when I clone the valid fields for submission, I loose the field values. That's the next thing to address.

It's still a bit unclear how all this will integrate with the real PayPal shopping cart. The cart shows on it's own page and so, we get returned to this page… but I've set things up using ajax which assumes we will never leave this page.