stable mid-point for diagram calculations
This gets all the basic values, margins, offsets, sizes of the divs…

Next is the positioning information.

To keep the visible steps (down and right) regular when the layers may be different sizes, this calc will locate the lower right point then derive the upper left point to hit it.

After that, how to set the css for the layers… and how to manage secondary layer components… and corner radiuses. Sigh.