expand out the diagram div structure and write css to control
things are coming together.

the db and the on-page div structure must both be extended to cover all cases. This gets part of that done.

Also, the part of the js code that writes the control rules to the page is added here.

Still using dummy color fills for visibility. Haven't decided on the hooks that will connect inputs with divs to set the proper background images.