clean up css and improve component setting
The css is being organized to manage the all the product configurations in an understandable way.

The pen loop and belt have been isolated so settings on one product diagram doesn't slop over onto the others

I've also attempted to get the liner drop to show up on Full Leather while slaving to the radio on Quarter Bound. Gonna have to debug that 'cause the radio isn't hiding it now.