lay html foundation for cumulative pricing display
add price and oldprice attributes on pertinent fields.

add span to fieldset legends (class 'plus') to show the amount added by the option.

I plan to add the total to the cart button.

bookbodies ordered with a reusable cover have a global json object on the page to do a look-up. Data defined in Catalog Model.

I've done some re-pricing of printed journals while making that lookup. But the process is not complete