work on bookbody-by-select-lists pricing system
Looks like I need a new property for 'caveat' accumulation; messages to show when user choices create uncertainty in the price.

These can be stored with each option and concatenated into a single message for output with the product price and description.

All this stuff can also be sent with all the emails

I could also look at adding each option as a shopping cart item