Install Product factory
Ordering products ranges from super-custom work with complex forms to simple purchases of items from stock.

The factory gets a utility to deal with the kind of product being ordered so the pricing can be calculated on the server (freeing me from dependence on values from the user).

The concrete products should be extended to do other tasks (yet to be defined) like making the array leaves for PayPal upload() cart calls.