partially strip productHelper factory
This fixes the bug the factory caused where summary/detail display data was not accumulated into a master list from the various concrete helpers.

Possibly the base class can become a composition of many concrete helpers instead. In this way, it can serve as the accumulator for the activity of all its components.