bring full menu onto page and establish basic css
to establish a full-depth cascading menu, all the data needed to be included on the page. This is done.

And css had to be added to show/hide the appropriate nodes. This is done.

But the li:hover simple strategy for revealing nests doesn't work because of overlapping LIs which cause the wrong ul nodes to activate.