establish a large chunk of the Quote flow
No detailed processes, like emailing a response or moving the project to Quote status... but the page to page pattern is being established.

The one (i think) hole is the submit button/form action for the initial page. It does not target the quote controller yet. but the receive a quote button does!